Add status request variable for flash partial (#101)
Actually, there is a {{ type }} variable for flash partial that display 'success' or 'danger' depending on the success or error of the ajax request.

It seems that "error" instead of "danger" would be a better name cause "danger" is bootstrap typed whereas "error" is the standard statusMessage returned by ajax.

To not break compatibility with old partial I added a "status" variable in addition to the type. Can help for partial customization.